servicio-ia/requirements.txt
# =============================================================================
# MedScribe AI - Servicio IA (Python)
# =============================================================================

# Web framework
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
httpx>=0.27.0

# LLM / Whisper (Groq es compatible con OpenAI SDK)
openai>=1.60.0
anthropic>=0.40.0

# Generacion de documentos
python-docx==1.1.0
reportlab==4.2.0

# Config
pydantic==2.9.0
python-dotenv==1.0.1

# Cola de trabajos asincrona + idempotencia + circuit breaker state
redis>=5.0.0
rq>=1.16.0

# Testing
pytest>=8.0.0

# RAG vectorial (embeddings + vector store)
qdrant-client>=1.11.0
sentence-transformers>=3.0.0

# Audio (necesario para diarizacion con Pyannote y conversion de audio)
numpy>=1.26,<2.0
soundfile==0.12.1
scipy>=1.11,<2.0

# Diarizacion local con Pyannote (Torch se instala aparte en el Dockerfile
# usando la rueda CPU-only para mantener la imagen pequena).
# Pin de huggingface_hub<1.0 porque pyannote 3.3.2 llama internamente a
# hf_hub_download con `use_auth_token=`, kwarg que huggingface_hub 1.0+
# elimino. pyannote 4.x lo arregla pero exige torch 2.8+ y torchcodec,
# lo que rompe la rueda CPU-only 2.5.1.
pyannote.audio==3.3.2
huggingface_hub<1.0
